Add some PR fixes
