--- src/jvm/test/resources/expected_logs/array_read_write_run_with_lambda.txt
@@ -1,26 +1,0 @@
-= Invalid execution results =
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-The following interleaving leads to the error:
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-Detailed trace:
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-|                                                                                   Thread 1                                                                                   |
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit                                                                                                                                          |
-|   failure$1.invoke() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:28)                              |
-|     ArrayReadWriteRunWithLambdaTest#1.operation() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:31) |
-|       IntArray#1[0].READ: 1 at ArrayReadWriteRunWithLambdaTest.operation(ArrayReadWriteRunWithLambdaTest.kt:22)                                                              |
-|       IntArray#1[0].WRITE(2) at ArrayReadWriteRunWithLambdaTest.operation(ArrayReadWriteRunWithLambdaTest.kt:22)                                                             |
-|       IntArray#1[0].READ: 2 at ArrayReadWriteRunWithLambdaTest.operation(ArrayReadWriteRunWithLambdaTest.kt:23)                                                              |
-|   result: kotlin.Unit                                                                                                                                                        |
-| ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
