+++ src/jvm/test/resources/expected_logs/array_rw_run_with_lambda.txt
@@ -1,0 +1,26 @@
+= Invalid execution results =
+| ----------------------------------- |
+|              Thread 1               |
+| ----------------------------------- |
+| run(() -> kotlin.Unit): kotlin.Unit |
+| ----------------------------------- |
+
+The following interleaving leads to the error:
+| ----------------------------------- |
+|              Thread 1               |
+| ----------------------------------- |
+| run(() -> kotlin.Unit): kotlin.Unit |
+| ----------------------------------- |
+
+Detailed trace:
+| --------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+|                                                                               Thread 1                                                                                |
+| --------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| run(() -> kotlin.Unit): kotlin.Unit                                                                                                                                   |
+|   failure$1.invoke() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:34)                          |
+|     ArrayReadWriteRunWithLambdaTest#1.block() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:37) |
+|       IntArray#1[0].READ: 1 at ArrayReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:49)                                                          |
+|       IntArray#1[0].WRITE(2) at ArrayReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:49)                                                         |
+|       IntArray#1[0].READ: 2 at ArrayReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:50)                                                          |
+|   result: kotlin.Unit                                                                                                                                                 |
+| --------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
