--- src/jvm/test/resources/expected_logs/atomic_references_name_two_fields_trace_run_with_lambda.txt
@@ -1,27 +1,0 @@
-= Invalid execution results =
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-The following interleaving leads to the error:
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-Detailed trace:
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-|                                                                                             Thread 1                                                                                             |
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| run(() -> kotlin.Unit): kotlin.Unit                                                                                                                                                              |
-|   failure$1.invoke() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:28)                                                  |
-|     AtomicReferencesFromMultipleFieldsRunWithLambdaTest#1.operation() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:31) |
-|       atomicReference1.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.operation(AtomicReferencesNamesRunWithLambdaTests.kt:103)                                  |
-|       atomicReference2.READ: AtomicReference#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.operation(AtomicReferencesNamesRunWithLambdaTests.kt:103)                                  |
-|       AtomicReference#1.get(): Node#1 at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.operation(AtomicReferencesNamesRunWithLambdaTests.kt:103)                                           |
-|       AtomicReference#1.compareAndSet(Node#1,Node#2): true at AtomicReferencesFromMultipleFieldsRunWithLambdaTest.operation(AtomicReferencesNamesRunWithLambdaTests.kt:103)                      |
-|   result: kotlin.Unit                                                                                                                                                                            |
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
