--- src/jvm/test/resources/expected_logs/atomic_references_names_trace_run_with_lambda.txt
@@ -1,52 +1,0 @@
-= Invalid execution results =
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-The following interleaving leads to the error:
-| ----------------------------------- |
-|              Thread 1               |
-| ----------------------------------- |
-| run(() -> kotlin.Unit): kotlin.Unit |
-| ----------------------------------- |
-
-Detailed trace:
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-|                                                                                       Thread 1                                                                                       |
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
-| run(() -> kotlin.Unit): kotlin.Unit                                                                                                                                                  |
-|   failure$1.invoke() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:28)                                      |
-|     AtomicReferencesNamesRunWithLambdaTests#1.operation() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(BaseRunWithLambdaRepresentationTest.kt:31) |
-|       AtomicReference#1.get(): Node#1 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:35)                                            |
-|       AtomicReference#1.compareAndSet(Node#1,Node#2): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:35)                       |
-|       AtomicReference#1.set(Node#3) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:36)                                              |
-|       AtomicInteger#1.get(): 3 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:38)                                                   |
-|       AtomicInteger#1.compareAndSet(3,2): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:38)                                   |
-|       AtomicInteger#1.set(3) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:39)                                                     |
-|       AtomicLong#1.get(): 3 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:41)                                                      |
-|       AtomicLong#1.compareAndSet(3,2): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:41)                                      |
-|       AtomicLong#1.set(3) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:42)                                                        |
-|       AtomicBoolean#1.get(): false at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:44)                                               |
-|       AtomicBoolean#1.compareAndSet(false,true): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:44)                            |
-|       AtomicBoolean#1.set(false) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:45)                                                 |
-|       AtomicReferenceArray#1[0].get(): Node#4 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:47)                                    |
-|       AtomicReferenceArray#1[0].compareAndSet(Node#4,Node#5): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:47)               |
-|       AtomicReferenceArray#1[0].set(Node#6) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:48)                                      |
-|       AtomicIntegerArray#1[0].get(): 2 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:50)                                           |
-|       AtomicIntegerArray#1[0].compareAndSet(2,1): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:50)                           |
-|       AtomicIntegerArray#1[0].set(2) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:51)                                             |
-|       AtomicLongArray#1[0].get(): 2 at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:53)                                              |
-|       AtomicLongArray#1[0].compareAndSet(2,1): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:53)                              |
-|       AtomicLongArray#1[0].set(2) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:54)                                                |
-|       AtomicReference#2.set(Node#7) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:56)                                              |
-|       AtomicIntegerArray#2[0].compareAndSet(1,2): false at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:57)                          |
-|       AtomicInteger#2.compareAndSet(0,2): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:59)                                   |
-|       AtomicInteger#2.set(0) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:60)                                                     |
-|       AtomicInteger#3.compareAndSet(1,2): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:62)                                   |
-|       AtomicInteger#3.set(3) at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:63)                                                     |
-|       AtomicIntegerArray#3[1].compareAndSet(0,1): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:65)                           |
-|       AtomicIntegerArray#4[1].compareAndSet(0,1): true at AtomicReferencesNamesRunWithLambdaTests.operation(AtomicReferencesNamesRunWithLambdaTests.kt:66)                           |
-|   result: kotlin.Unit                                                                                                                                                                |
-| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
