+++ src/jvm/test/resources/expected_logs/var_rw_run_with_lambda.txt
@@ -1,0 +1,28 @@
+= Invalid execution results =
+| ----------------------------------- |
+|              Thread 1               |
+| ----------------------------------- |
+| run(() -> kotlin.Unit): kotlin.Unit |
+| ----------------------------------- |
+
+The following interleaving leads to the error:
+| ----------------------------------- |
+|              Thread 1               |
+| ----------------------------------- |
+| run(() -> kotlin.Unit): kotlin.Unit |
+| ----------------------------------- |
+
+Detailed trace:
+| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+|                                                                                 Thread 1                                                                                 |
+| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
+| run(() -> kotlin.Unit): kotlin.Unit                                                                                                                                      |
+|   failure$1.invoke() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:34)                             |
+|     VariableReadWriteRunWithLambdaTest#1.block() at BaseRunWithLambdaRepresentationTest$testRunWithModelChecker$failure$1.invoke(RunWithLambdaRepresentationTests.kt:37) |
+|       x.READ: 0 at VariableReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:149)                                                                     |
+|       x.WRITE(1) at VariableReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:149)                                                                    |
+|       x.READ: 1 at VariableReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:150)                                                                     |
+|       x.WRITE(0) at VariableReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:150)                                                                    |
+|       x.READ: 0 at VariableReadWriteRunWithLambdaTest.block(RunWithLambdaRepresentationTests.kt:150)                                                                     |
+|   result: kotlin.Unit                                                                                                                                                    |
+| ------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
